moved schemas tests into correct folder area, should exclude them from results
