+++ test/schemas/components/email.test.js
@@ -1,13 +1,13 @@
-const { TEMPORARY, PERMANENT } = require('../../app/constants/notify-simulation-email-addresses')
+const { TEMPORARY, PERMANENT } = require('../../../app/constants/notify-simulation-email-addresses')
 
-const schema = require('../../app/schemas/components/email-lenient')
+const schema = require('../../../app/schemas/components/email')
 
 let email
 
-describe('Lenient email schema', () => {
+describe('Email schema', () => {
   describe('When email is a valid email address', () => {
     beforeEach(() => {
-      email = require('../../test/mocks/components/email')
+      email = require('../../mocks/components/email')
     })
 
     test('should return an object', async () => {
@@ -263,24 +263,29 @@ describe('Lenient email schema', () => {
       expect(result).toBeInstanceOf(Object)
     })
 
-    test('should return an object with 1 key', () => {
+    test('should return an object with 2 keys', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toHaveLength(1)
+      expect(Object.keys(result)).toHaveLength(2)
     })
 
-    test('should return an object with keys: value', () => {
+    test('should return an object with keys: value, error', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toStrictEqual(['value'])
+      expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toBe(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should not return an object with key: error', () => {
-      const result = schema.validate(email)
-      expect(Object.keys(result)).not.toContain('error')
+    test('should return "any.required" as value for key: error.details[0].type', () => {
+      const result = schema.validate()
+      expect(result.error.details[0].type).toBe('any.required')
+    })
+
+    test('should return "Email must be provided." as value for key: error.details[0].message', () => {
+      const result = schema.validate()
+      expect(result.error.details[0].message).toBe('Email must be provided.')
     })
   })
 
@@ -294,24 +299,29 @@ describe('Lenient email schema', () => {
       expect(result).toBeInstanceOf(Object)
     })
 
-    test('should return an object with 1 key', () => {
+    test('should return an object with 2 keys', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toHaveLength(1)
+      expect(Object.keys(result)).toHaveLength(2)
     })
 
-    test('should return an object with keys: value', () => {
+    test('should return an object with keys: value, error', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toStrictEqual(['value'])
+      expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toBe(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should not return an object with key: error', () => {
-      const result = schema.validate(email)
-      expect(Object.keys(result)).not.toContain('error')
+    test('should return "any.required" as value for key: error.details[0].type', () => {
+      const result = schema.validate()
+      expect(result.error.details[0].type).toBe('any.required')
+    })
+
+    test('should return "Email must be provided." as value for key: error.details[0].message', () => {
+      const result = schema.validate()
+      expect(result.error.details[0].message).toBe('Email must be provided.')
     })
   })
 
@@ -325,24 +335,29 @@ describe('Lenient email schema', () => {
       expect(result).toBeInstanceOf(Object)
     })
 
-    test('should return an object with 1 key', () => {
+    test('should return an object with 2 keys', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toHaveLength(1)
+      expect(Object.keys(result)).toHaveLength(2)
     })
 
-    test('should return an object with keys: value', () => {
+    test('should return an object with keys: value, error', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toStrictEqual(['value'])
+      expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toBe(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should not return an object with key: error', () => {
-      const result = schema.validate(email)
-      expect(Object.keys(result)).not.toContain('error')
+    test('should return "any.required" as value for key: error.details[0].type', () => {
+      const result = schema.validate()
+      expect(result.error.details[0].type).toBe('any.required')
+    })
+
+    test('should return "Email must be provided." as value for key: error.details[0].message', () => {
+      const result = schema.validate()
+      expect(result.error.details[0].message).toBe('Email must be provided.')
     })
   })
 
@@ -356,24 +371,29 @@ describe('Lenient email schema', () => {
       expect(result).toBeInstanceOf(Object)
     })
 
-    test('should return an object with 1 key', () => {
+    test('should return an object with 2 keys', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toHaveLength(1)
+      expect(Object.keys(result)).toHaveLength(2)
     })
 
-    test('should return an object with keys: value', () => {
+    test('should return an object with keys: value, error', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toStrictEqual(['value'])
+      expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toBe(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should not return an object with key: error', () => {
-      const result = schema.validate(email)
-      expect(Object.keys(result)).not.toContain('error')
+    test('should return "any.required" as value for key: error.details[0].type', () => {
+      const result = schema.validate()
+      expect(result.error.details[0].type).toBe('any.required')
+    })
+
+    test('should return "Email must be provided." as value for key: error.details[0].message', () => {
+      const result = schema.validate()
+      expect(result.error.details[0].message).toBe('Email must be provided.')
     })
   })
 
@@ -387,24 +407,29 @@ describe('Lenient email schema', () => {
       expect(result).toBeInstanceOf(Object)
     })
 
-    test('should return an object with 1 key', () => {
+    test('should return an object with 2 keys', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toHaveLength(1)
+      expect(Object.keys(result)).toHaveLength(2)
     })
 
-    test('should return an object with keys: value', () => {
+    test('should return an object with keys: value, error', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toStrictEqual(['value'])
+      expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toBe(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should not return an object with key: error', () => {
-      const result = schema.validate(email)
-      expect(Object.keys(result)).not.toContain('error')
+    test('should return "any.required" as value for key: error.details[0].type', () => {
+      const result = schema.validate()
+      expect(result.error.details[0].type).toBe('any.required')
+    })
+
+    test('should return "Email must be provided." as value for key: error.details[0].message', () => {
+      const result = schema.validate()
+      expect(result.error.details[0].message).toBe('Email must be provided.')
     })
   })
 
@@ -418,24 +443,29 @@ describe('Lenient email schema', () => {
       expect(result).toBeInstanceOf(Object)
     })
 
-    test('should return an object with 1 key', () => {
+    test('should return an object with 2 keys', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toHaveLength(1)
+      expect(Object.keys(result)).toHaveLength(2)
     })
 
-    test('should return an object with keys: value', () => {
+    test('should return an object with keys: error and value', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toStrictEqual(['value'])
+      expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toStrictEqual(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should not return an object with key: error', () => {
+    test('should return "string.email" as value for key: error.details[0].type', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).not.toContain('error')
+      expect(result.error.details[0].type).toBe('string.email')
+    })
+
+    test('should return "The email provided is invalid." as value for key: error.details[0].message', () => {
+      const result = schema.validate(email)
+      expect(result.error.details[0].message).toBe('The email provided is invalid.')
     })
   })
 
@@ -449,24 +479,29 @@ describe('Lenient email schema', () => {
       expect(result).toBeInstanceOf(Object)
     })
 
-    test('should return an object with 1 key', () => {
+    test('should return an object with 2 keys', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toHaveLength(1)
+      expect(Object.keys(result)).toHaveLength(2)
     })
 
-    test('should return an object with keys: value', () => {
+    test('should return an object with keys: error and value', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toStrictEqual(['value'])
+      expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toStrictEqual(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should not return an object with key: error', () => {
+    test('should return "string.email" as value for key: error.details[0].type', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).not.toContain('error')
+      expect(result.error.details[0].type).toBe('string.email')
+    })
+
+    test('should return "The email provided is invalid." as value for key: error.details[0].message', () => {
+      const result = schema.validate(email)
+      expect(result.error.details[0].message).toBe('The email provided is invalid.')
     })
   })
 
@@ -480,24 +515,29 @@ describe('Lenient email schema', () => {
       expect(result).toBeInstanceOf(Object)
     })
 
-    test('should return an object with 1 key', () => {
+    test('should return an object with 2 keys', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toHaveLength(1)
+      expect(Object.keys(result)).toHaveLength(2)
     })
 
-    test('should return an object with keys: value', () => {
+    test('should return an object with keys: error and value', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toStrictEqual(['value'])
+      expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toStrictEqual(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should not return an object with key: error', () => {
+    test('should return "string.email" as value for key: error.details[0].type', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).not.toContain('error')
+      expect(result.error.details[0].type).toBe('string.email')
+    })
+
+    test('should return "The email provided is invalid." as value for key: error.details[0].message', () => {
+      const result = schema.validate(email)
+      expect(result.error.details[0].message).toBe('The email provided is invalid.')
     })
   })
 
@@ -523,12 +563,7 @@ describe('Lenient email schema', () => {
 
     test('should return email as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toStrictEqual(email)
-    })
-
-    test('should not return an object with key: error', () => {
-      const result = schema.validate(email)
-      expect(Object.keys(result)).not.toContain('error')
+      expect(result.value).toBe(email)
     })
   })
 
@@ -542,24 +577,29 @@ describe('Lenient email schema', () => {
       expect(result).toBeInstanceOf(Object)
     })
 
-    test('should return an object with 1 key', () => {
+    test('should return an object with 2 keys', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toHaveLength(1)
+      expect(Object.keys(result)).toHaveLength(2)
     })
 
-    test('should return an object with keys: value', () => {
+    test('should return an object with keys: error and value', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toStrictEqual(['value'])
+      expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toStrictEqual(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should not return an object with key: error', () => {
+    test('should return "string.email" as value for key: error.details[0].type', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).not.toContain('error')
+      expect(result.error.details[0].type).toBe('string.email')
+    })
+
+    test('should return "The email provided is invalid." as value for key: error.details[0].message', () => {
+      const result = schema.validate(email)
+      expect(result.error.details[0].message).toBe('The email provided is invalid.')
     })
   })
 
@@ -573,24 +613,29 @@ describe('Lenient email schema', () => {
       expect(result).toBeInstanceOf(Object)
     })
 
-    test('should return an object with 1 key', () => {
+    test('should return an object with 2 keys', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toHaveLength(1)
+      expect(Object.keys(result)).toHaveLength(2)
     })
 
-    test('should return an object with keys: value', () => {
+    test('should return an object with keys: error and value', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toStrictEqual(['value'])
+      expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toStrictEqual(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should not return an object with key: error', () => {
+    test('should return "string.email" as value for key: error.details[0].type', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).not.toContain('error')
+      expect(result.error.details[0].type).toBe('string.email')
+    })
+
+    test('should return "The email provided is invalid." as value for key: error.details[0].message', () => {
+      const result = schema.validate(email)
+      expect(result.error.details[0].message).toBe('The email provided is invalid.')
     })
   })
 
@@ -636,24 +681,29 @@ describe('Lenient email schema', () => {
       expect(result).toBeInstanceOf(Object)
     })
 
-    test('should return an object with 1 key', () => {
+    test('should return an object with 2 keys', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toHaveLength(1)
+      expect(Object.keys(result)).toHaveLength(2)
     })
 
-    test('should return an object with keys: value', () => {
+    test('should return an object with keys: error and value', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).toStrictEqual(['value'])
+      expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toStrictEqual(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should not return an object with key: error', () => {
+    test('should return "string.empty" as value for key: error.details[0].type', () => {
       const result = schema.validate(email)
-      expect(Object.keys(result)).not.toContain('error')
+      expect(result.error.details[0].type).toBe('string.empty')
+    })
+
+    test('should return "Email cannot be empty." as value for key: error.details[0].message', () => {
+      const result = schema.validate(email)
+      expect(result.error.details[0].message).toBe('Email cannot be empty.')
     })
   })
 
@@ -677,14 +727,14 @@ describe('Lenient email schema', () => {
       expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toBe(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should return "string.base" as value for key: error.details[0].type', () => {
+    test('should return "alternatives.types" as value for key: error.details[0].type', () => {
       const result = schema.validate(email)
-      expect(result.error.details[0].type).toBe('string.base')
+      expect(result.error.details[0].type).toBe('alternatives.types')
     })
 
     test('should return "Email must be a string." as value for key: error.details[0].message', () => {
@@ -713,14 +763,14 @@ describe('Lenient email schema', () => {
       expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toBe(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should return "string.base" as value for key: error.details[0].type', () => {
+    test('should return "alternatives.types" as value for key: error.details[0].type', () => {
       const result = schema.validate(email)
-      expect(result.error.details[0].type).toBe('string.base')
+      expect(result.error.details[0].type).toBe('alternatives.types')
     })
 
     test('should return "Email must be a string." as value for key: error.details[0].message', () => {
@@ -749,14 +799,14 @@ describe('Lenient email schema', () => {
       expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toBe(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should return "string.base" as value for key: error.details[0].type', () => {
+    test('should return "alternatives.types" as value for key: error.details[0].type', () => {
       const result = schema.validate(email)
-      expect(result.error.details[0].type).toBe('string.base')
+      expect(result.error.details[0].type).toBe('alternatives.types')
     })
 
     test('should return "Email must be a string." as value for key: error.details[0].message', () => {
@@ -785,14 +835,14 @@ describe('Lenient email schema', () => {
       expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toBe(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should return "string.base" as value for key: error.details[0].type', () => {
+    test('should return "alternatives.types" as value for key: error.details[0].type', () => {
       const result = schema.validate(email)
-      expect(result.error.details[0].type).toBe('string.base')
+      expect(result.error.details[0].type).toBe('alternatives.types')
     })
 
     test('should return "Email must be a string." as value for key: error.details[0].message', () => {
@@ -821,9 +871,9 @@ describe('Lenient email schema', () => {
       expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toBe(email)
+      expect(result.value).toBeUndefined()
     })
 
     test('should return "any.required" as value for key: error.details[0].type', () => {
@@ -857,14 +907,14 @@ describe('Lenient email schema', () => {
       expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toBe(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should return "string.base" as value for key: error.details[0].type', () => {
+    test('should return "alternatives.types" as value for key: error.details[0].type', () => {
       const result = schema.validate(email)
-      expect(result.error.details[0].type).toBe('string.base')
+      expect(result.error.details[0].type).toBe('alternatives.types')
     })
 
     test('should return "Email must be a string." as value for key: error.details[0].message', () => {
@@ -893,14 +943,14 @@ describe('Lenient email schema', () => {
       expect(Object.keys(result)).toStrictEqual(['value', 'error'])
     })
 
-    test('should return email as value for key: value', () => {
+    test('should return undefined as value for key: value', () => {
       const result = schema.validate(email)
-      expect(result.value).toBe(email)
+      expect(result.value).toBeUndefined()
     })
 
-    test('should return "string.base" as value for key: error.details[0].type', () => {
+    test('should return "alternatives.types" as value for key: error.details[0].type', () => {
       const result = schema.validate(email)
-      expect(result.error.details[0].type).toBe('string.base')
+      expect(result.error.details[0].type).toBe('alternatives.types')
     })
 
     test('should return "Email must be a string." as value for key: error.details[0].message', () => {
